requirements.txt
flask
scikit-learn
numpy
pillow
torch
torchvision